+++ ps_stack.h
@@ -6,7 +6,7 @@
 /*   By: ny-handr <ny-handr@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/10 12:51:13 by todina-r          #+#    #+#             */
-/*   Updated: 2026/03/11 12:09:27 by ny-handr         ###   ########.fr       */
+/*   Updated: 2026/03/27 09:21:54 by todina-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 typedef struct s_st_node
 {
-	int					value;
+	long				value;
 	struct s_st_node	*next;
 }				t_st_node;
 
@@ -28,7 +28,7 @@ typedef struct s_stack
 	size_t		size;
 }				t_stack;
 
-t_st_node	*st_new(int val);
+t_st_node	*st_new(long val);
 t_st_node	*st_second_last(t_stack stack);
 void		st_add_front(t_stack *stack, t_st_node *node);
 void		st_add_back(t_stack *stack, t_st_node *node);
